# Receive: Dont rebatch already replicated requests