@matej-g 2022-11-29
Closing as this has been superseded by https://github.com/thanos-io/thanos/pull/5910